resolve: Turn `resolve_error` into a method on `Resolver`
Rename it to `report_error` and move into `diagnostics.rs`

Also turn `check_unused` into a method on `Resolver`
